Add well-known genesisData in config (#4480)
* Add genesisData for known networks

* Use genesisData from known networks

* Remove duplicated networkGenesis data in lightclient
