+++ packages/light-client/src/index.ts
@@ -21,7 +21,6 @@ import {
   activeHeader,
   assertValidFinalityProof,
 } from "./validation.js";
-import {GenesisData} from "./networks.js";
 import {getLcLoggerConsole, ILcLogger} from "./utils/logger.js";
 import {computeSyncPeriodAtEpoch, computeSyncPeriodAtSlot, computeEpochAtSlot} from "./utils/clock.js";
 
@@ -29,13 +28,15 @@ import {computeSyncPeriodAtEpoch, computeSyncPeriodAtSlot, computeEpochAtSlot} f
 export {LightclientEvent} from "./events.js";
 export {SyncCommitteeFast} from "./types.js";
 
+export type GenesisData = {
+  genesisTime: number;
+  genesisValidatorsRoot: RootHex | Uint8Array;
+};
+
 export type LightclientInitArgs = {
   config: IChainForkConfig;
   logger?: ILcLogger;
-  genesisData: {
-    genesisTime: number;
-    genesisValidatorsRoot: RootHex | Uint8Array;
-  };
+  genesisData: GenesisData;
   beaconApiUrl: string;
   snapshot: {
     header: phase0.BeaconBlockHeader;
